+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 #### Next Release
 - Added Query.first()
+- Added Query.firstOrDefault()
 
 #### v0.1.0 (2019-11-22)
 - Initial release